Rename patch files
Now that we're using 'gbp pq' to manage patches, things like
shuffling them around becomes much more convenient and thus will
certainly happen more frequently; with that in mind, it's better
to drop the prefix now and avoid further churn down the line.

Debian-specific patches, if we ever end up needing them, will
either use debian: as a prefix or be stored in a debian/
subdirectory, so it will still be easy to quickly tell them apart
from those that are intended for upstream.

Since we're renaming things anyway, we can let 'gbp pq' pick the
names of the patch files for us from now on.